// src/types.ts
export type CompanyId = 'pallyra' | 'guennita' | 'mimada';

export interface Product {
  id: string;
  code: string;
  company: CompanyId;
  product_name: string;
  description: string;
  isWholesaleEnabled?: boolean;
  retail_price: number;
  wholesale_price: number;
  wholesale_min_qty: number;
  wholesale_max_qty?: number;
  original_price: number;
  current_price: number;
  image: string;
  image_hover?: string;
  images?: string[];
  category: string;
  subcategory: string;
  isVisible: boolean;
  isFeatured: boolean;
  insumos?: { insumoId: string; quantity: number }[];
  variations?: Variation[];
  estimatedCost?: number;
  activeInCatalog?: boolean;
  salesCount?: number;
  clicksCount?: number;
  isLastUnits?: boolean;
  createdAt?: any;
  giftInfo?: string;
  imageSettings?: {
    scale?: number;
    translateX?: number;
    translateY?: number;
    rotate?: number;
  };
}

export interface Variation {
  id: string;
  name: string; // Cor, Tamanho, etc
  type: 'single' | 'multiple';
  options: {
    name: string;
    price: number;
    stockId?: string;
  }[];
}
// ...
export interface CartItem extends Product {
  quantity: number;
  productId?: string;
}